splitting of "marqo_external_utils" into smaller sub modules
